fix/chunksize-by-filesize

+++ include/placeholders_interface/Placeholders.h
@@ -6,7 +6,7 @@ public:
     static void CreateOne(
         _In_ PCWSTR fileName,
         _In_ PCWSTR fileIdentity,
-        uint32_t fileSize,
+        int64_t fileSize,
         DWORD fileIdentityLength,
         uint32_t fileAttributes,
         FILETIME creationTime,
